PROGRAMMERS HATE THIS SIMPLE TRICK
https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/operators/arithmetic-operators#integer-division

+++ Content.Server/Light/EntitySystems/PoweredLightSystem.cs
@@ -277,7 +277,7 @@ namespace Content.Server.Light.EntitySystems
                         if (time > light.LastThunk + ThunkDelay)
                         {
                             var x = ((_random.Next() % 3) + (_random.Next() % 3)) / 2; // major seconds
-                            var y = _random.Next(0, 10) / 10; // minor, so that our lamps will turn on more randomly
+                            var y = (float)_random.Next(0, 10) / 10; // minor, so that our lamps will turn on more randomly
                             var f = x + y + 0.3; // major + minor + basetime
                             _appearance.SetData(uid, PoweredLightVisuals.BulbState, PoweredLightState.On, appearance);
                             uid.SpawnTimer(TimeSpan.FromSeconds(f), () =>
